DRY partial validations into concern, add spec that they're being used
