+++ app/models/liquid_layout.rb
@@ -1,23 +1,9 @@
 class LiquidLayout < ActiveRecord::Base
+  include HasLiquidPartials
+
   has_many :pages
   validates :title, presence: true, allow_blank: false
   validates :content, presence: true, allow_blank: false
-
-  validate :no_unknown_partials
-
-  def partial_names
-    LiquidTagFinder.new(content).partial_names
-  end
-
-  def partial_refs
-    LiquidTagFinder.new(content).partial_refs
-  end
-
-  def no_unknown_partials
-    LiquidPartial.missing_partials(partial_names).each do |name|
-      errors.add :content, "includes unknown partial '#{name}'"
-    end
-  end
 
   class << self
     def default
