+++ app/models/liquid_partial.rb
@@ -1,13 +1,10 @@
 class LiquidPartial < ActiveRecord::Base
+  include HasLiquidPartials
+
   validates :title, presence: true, allow_blank: false
   validates :content, presence: true, allow_blank: false
 
   validate :one_plugin
-  validate :no_unknown_partials
-
-  def partial_names
-    LiquidTagFinder.new(content).partial_names
-  end
 
   def plugin_name
     LiquidTagFinder.new(content).plugin_names[0]
@@ -26,22 +23,18 @@ class LiquidPartial < ActiveRecord::Base
     collector
   end
 
+  # Filters array of partial names to those absent from the database. (returns new array)
+  def self.missing_partials(names)
+    names.reject{|name| LiquidPartial.exists?(title: name) }
+  end
+
+  private
+
   def one_plugin
     plugin_names = LiquidTagFinder.new(content).plugin_names
     if plugin_names.size > 1
       errors.add(:content, "can only reference one partial, but found #{plugin_names.join(',')}")
     end
-  end
-
-  def no_unknown_partials
-    LiquidPartial.missing_partials(partial_names).each do |name|
-      errors.add :content, "includes unknown partial '#{name}'"
-    end
-  end
-
-  # Filters array of partial names to those absent from the database. (returns new array)
-  def self.missing_partials(names)
-    names.reject{|name| LiquidPartial.exists?(title: name) }
   end
 
 end
